requirements.txt
easydict==1.10
einops==0.6.0
h5py==2.10.0
numpy==1.22.3
PyYAML==6.0
torch==1.10.0+cu113
tqdm==4.64.1
transformers==4.27.1
